+++ Scripts/Mushroom.cs
@@ -18,7 +18,7 @@ public class Mushroom : MonoBehaviour {
 	public float maxScale = 1;
 	void UpdateScale () {
 
-		float scaleAmount = Mathf.InverseLerp(0.24f, 0.85f, SunControl.instance.posInDay);
+		float scaleAmount = Mathf.InverseLerp(0.24f, 0.85f, (float)SunControl.instance.posInDay);
 		float minScale = this.minScale != 0 ? this.minScale/100 : initScale;
 		float currentScalar = Mathf.Lerp (minScale, maxScale/100, scaleAmount);
 		transform.localScale = Vector3.one * currentScalar;
